requirements.txt
apiview==1.3.25
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
backports.zoneinfo==0.2.1
bootstrap4==0.1.0
certifi @ file:///private/var/folders/nz/j6p8yfhx1mv_0grj5xl4650h0000gp/T/abs_884c889c-96af-444f-bd6d-daddb5e9a462ykj3l5n_/croots/recipe/certifi_1655968814730/work/certifi
click==8.1.3
cycler==0.11.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1
django-bootstrap5==22.1
django-import-export==2.8.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
fonttools==4.36.0
gensim==4.2.0
h11==0.13.0
idna==3.3
joblib==1.1.0
JPype1==1.4.0
kiwisolver==1.4.4
konlpy==0.6.0
lxml==4.9.1
MarkupPy==1.14
matplotlib==3.5.3
nltk==3.7
numpy==1.23.2
odfpy==1.4.1
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
regex==2022.8.17
scikit-learn==1.1.2
scipy==1.9.0
selenium==4.4.3
six==1.16.0
smart-open==6.1.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
tablib==3.2.1
threadpoolctl==3.1.0
tqdm==4.64.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.12
wordcloud==1.8.2.2
wsproto==1.1.0
xlrd==2.0.1
xlwt==1.3.0
